Large changes made...
Figured out the basic structure for the purchase of an upgrade. These changes can be seen in the render function of the main game state in StateFactory and in the TradingPost Object.

This needs some heavy refactoring however due to the heavy reliance of copy and past to extend the functionality to the other game entities.

I think the way i will do this is through making a static class that will take a game entitiy and a kingdom object and spit out an object of the required resources. this will then be used to calculate if the player has the required resources.

I will also make a static method to handle the figuring out if the player has enough resources to make a purchase. These changes will be appended to my class diagram (physical copy)

There is also a definite scalabiltiy issue in the render function of the main game state. It is very intensive and unneeded for the browser to make mutiple if calls in a render function. I will need to look into a solution for this.

I also need to make a way to load and save the game states. Implementing this earlier will save time rejigging the system to work after implementing everything else.