Hooked up the deobfuscaor/obfusactor with the load/save methods . this just needs to be implemented in the view.
I also changed the internal workings of Building.js I decided to use the get set keywords to shorten the code base and make the building class more readable.

Also removed the word building from the unlockBuilding and buildBuilding methods. I decided that future readers of the code (me) would understand that the method build in the class building would probably refer to building the building it is a method of and that i didnt need to specify this in the method names as well.

upgrade will only increment the level by a given upgrade amount. I need to make a few calculator classes for the cost and gain of an upgrade transaction.